try to make stuff work
